2012-09-18 (tinybit)accepted LZMAed preset-menu. font lines can be appended to both preset-menu and menu.lst. 2012-09-08 (tinybit)added new map options --skip-sectors and --max-sectors. 2012-08-23 (tinybit)added variables to identify the CPU with cpuid. 2012-08-17 (tinybit)removed setkey support for Control, Shift, Alt and CapsLock keys. removed HLT instructions to satisfy buggy AMDs. getkey and chechkey functions were re-written to avoid using the int16h BIOS call. added chainloader support for bootmgr inside exFAT.
git-svn-id: https://grub4dos-chenall.googlecode.com/svn/branches/yaya-0.4.6a@304 ac2b6734-dafe-11de-aa1f-bbd297b2055c

+++ grub/asmstub.c
@@ -345,7 +345,7 @@ unset_int15_handler (void)
 
 /* The key map.  */
 unsigned short bios_key_map[KEY_MAP_SIZE + 1];
-unsigned short ascii_key_map[KEY_MAP_SIZE + 1];
+unsigned long ascii_key_map[KEY_MAP_SIZE + 1];
 
 int
 get_code_end (void)
