+++ grub/main.c
@@ -35,7 +35,7 @@ int grub_stage2 (void);
 
 char *program_name = 0;
 //int use_config_file = 1;
-int use_preset_menu = 0;
+unsigned long use_preset_menu = 0;
 #ifdef HAVE_LIBCURSES
 int use_curses = 1;
 #else
